Feat: Atoms/Tooltip replace custom tooltip with tippyjs Build: Moving fortawesome to bundled dependencies
