+++ package.json
@@ -42,13 +42,16 @@
   "module": "dist/index.esm.js",
   "types": "dist/index.d.ts",
   "style": "disc/style.min.css",
+  "bundledDependencies": [
+    "@fortawesome/pro-light-svg-icons",
+    "@fortawesome/pro-regular-svg-icons",
+    "@fortawesome/pro-solid-svg-icons"
+  ],
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^6.1.1",
     "@fortawesome/react-fontawesome": "^0.1.18",
-    "@fortawesome/pro-light-svg-icons": "^6.1.1",
-    "@fortawesome/pro-regular-svg-icons": "^6.1.1",
-    "@fortawesome/pro-solid-svg-icons": "^6.1.1",
     "@szhsin/react-menu": "^3.0.0",
+    "@tippyjs/react": "^4.2.6",
     "classnames": "^2.3.1",
     "color-alpha": "^1.1.3",
     "color-rgba": "^2.4.0",
@@ -56,8 +59,8 @@
     "lodash": "^4.17.21",
     "numeral": "^2.0.6",
     "react-datepicker": "^4.7.0",
-    "react-router-dom": "^6.2.2",
     "react-number-format": "^4.9.1",
+    "react-router-dom": "^6.2.2",
     "react-select": "^5.3.0",
     "suneditor": "^2.42.0",
     "suneditor-react": "^3.4.0",
@@ -69,6 +72,9 @@
   },
   "devDependencies": {
     "@babel/core": "^7.17.9",
+    "@fortawesome/pro-light-svg-icons": "^6.1.1",
+    "@fortawesome/pro-regular-svg-icons": "^6.1.1",
+    "@fortawesome/pro-solid-svg-icons": "^6.1.1",
     "@rollup/plugin-commonjs": "^21.0.3",
     "@rollup/plugin-node-resolve": "^13.2.0",
     "@storybook/addon-a11y": "^6.4.21",
@@ -118,5 +124,10 @@
   },
   "engines": {
     "node": ">=14.0.0"
-  }
+  },
+  "bundleDependencies": [
+    "@fortawesome/pro-light-svg-icons",
+    "@fortawesome/pro-regular-svg-icons",
+    "@fortawesome/pro-solid-svg-icons"
+  ]
 }
