+++ src/Components/Atoms/Tooltip/__test__/Tooltip.test.tsx
@@ -1,49 +1,34 @@
 import React from 'react';
-import { act, render, screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { render } from '@testing-library/react';
 
 import Tooltip from '../Tooltip';
-import { MenuDirectionEnum } from '../tooltipUtils';
+import { MenuDirectionEnum } from '../types';
 
 describe('Tooltip Component', () => {
   it('Tooltip renders', () => {
-    const { container } = render(<Tooltip tooltip='TOOLTIP'>CHILDREN</Tooltip>);
-    expect(container).toMatchSnapshot();
-  });
-
-  it('Tooltip renders with non default props', () => {
     const { container } = render(
-      <Tooltip
-        arrow={false}
-        className={'CLASSNAME'}
-        delay={100}
-        direction={MenuDirectionEnum.RIGHT}
-        extraMargin={100}
-        maxWidth={20}
-        style={{ height: '50px' }}
-        tooltip='TOOLTIP'>
-        CHILDREN
+      <Tooltip tooltip='TOOLTIP'>
+        <>CHILDREN</>
       </Tooltip>,
     );
     expect(container).toMatchSnapshot();
   });
 
-  it('Tooltip renders shows and closes', async () => {
-    const { container } = render(<Tooltip tooltip='TOOLTIP'>CHILDREN</Tooltip>);
+  it('Tooltip renders with empty tooltip', () => {
+    const { container } = render(
+      <Tooltip arrow={false} delay={100} direction={MenuDirectionEnum.RIGHT}>
+        <>CHILDREN</>
+      </Tooltip>,
+    );
     expect(container).toMatchSnapshot();
+  });
 
-    const tooltip = await screen.findByText('CHILDREN');
-
-    act(() => {
-      userEvent.hover(tooltip);
-    });
-
-    expect(container).toMatchSnapshot();
-
-    act(() => {
-      userEvent.click(tooltip);
-    });
-
+  it('Tooltip renders disabled', () => {
+    const { container } = render(
+      <Tooltip tooltip='TOOLTIP' disabled>
+        <>CHILDREN</>
+      </Tooltip>,
+    );
     expect(container).toMatchSnapshot();
   });
 });
