+++ src/Components/Atoms/Tooltip/types.ts
@@ -1,0 +1,6 @@
+export enum MenuDirectionEnum {
+  LEFT = 'left',
+  RIGHT = 'right',
+  TOP = 'top',
+  BOTTOM = 'bottom',
+}
