+++ src/assets/_tooltip.scss
@@ -1,85 +1,1 @@
-.gds-tooltip {
-  position: fixed;
-  text-shadow: none;
-  text-transform: none;
-  white-space: normal;
-  word-break: normal;
-  word-spacing: normal;
-  word-wrap: normal;
-  display: inline-block;
-  z-index: 10;
-  &.right {
-    margin-left: 5px;
-
-    .tooltip-arrow {
-      top: 50%;
-      left: auto;
-      margin-left: -5px;
-      border-width: 5px 5px 5px 0;
-      border-right-color: $charcoal;
-    }
-  }
-
-  &.left {
-    margin-left: -5px;
-
-    .tooltip-arrow {
-      top: 50%;
-      margin-top: -5px;
-      border-width: 5px 0 5px 5px;
-      border-left-color: $charcoal;
-      right: -5px;
-      left: auto;
-    }
-  }
-
-  &.top {
-    margin-top: -5px;
-
-    .tooltip-arrow {
-      top: auto;
-      bottom: -5px;
-      left: 50%;
-      margin-left: -5px;
-      border-width: 5px 5px 0;
-      border-top-color: $charcoal;
-    }
-  }
-
-  &.bottom {
-    margin-top: 5px;
-
-    .tooltip-arrow {
-      top: 0;
-      left: 50%;
-      margin-left: -5px;
-      border-width: 0 5px 5px;
-      border-bottom-color: $charcoal;
-    }
-  }
-
-  .tooltip-arrow {
-    top: 50%;
-    left: 0;
-    margin-top: -5px;
-    border-width: 5px 5px 5px 0;
-    border-right-color: $charcoal;
-    position: absolute;
-    width: 0;
-    height: 0;
-    border-color: transparent;
-    border-right-color: transparent;
-    border-style: solid;
-  }
-
-  .tooltip-label {
-    max-width: 300px;
-    padding: 6px 8px;
-    color: $white;
-    text-align: center;
-    background-color: $charcoal;
-    border-radius: 4px;
-    overflow-x: hidden;
-    white-space: wrap;
-  }
-}
+@import '~tippy.js/dist/tippy.css';
